gdb: add or1k and or1knd target support
This patch is pretty much a copy of the current
GDB 7.2 port of the openrisc processor.

It now supports or1knd (no delay slot target).
This architecture has not been tested.
The default target is or1k (with delay slot).
You can change the target arch with:

(gdb) set architecture or1knd
The target architecture is assumed to be or1knd

Signed-off-by: Franck Jullien <franck.jullien@gmail.com>

gdb/ChangeLog:

	* or1k-tdesc.c  : Add file for or1k support
	* or1k-tdesc.h  : Add file for or1k support
	* configure.tgt : Add or1k targets
	* gdb.texinfo   : Add OpenRISC documentation