or1ksim: Fix issue with disable-sim still linking or1ksim
Pointed out by travis. When building with --disable-sim the build
still trys to link in libsim from or1ksim unless we explicitly pass
disable-or1ksim.

This change fixes this behavior, but its a bit complicated because the
check for or1ksim and enable_sim are in completely different places.
What we do here is first set the or1ksim ldflags in the temporary
variable then later if sim is indeed enables we append the ldflags into
the build.