travis: Doesnt need guile, also make config more clear
The current config line says --disable-or1ksim but its a bit unclear. We
are actually enabling sim (or1k) but disabiling the or1ksim wrapper.
Add a --enable-sim line to be a bit more clear that sim is enabled.
Also remove guile-1.8 dependency since thats only needed for generating
the cgen stuff which is already generated.