sim: testsuite: Get common tests compiling again
This is a series of changes to the the old and non working sim common
testsuite .  As far as I can tell this has not been used since what
looks like before 64-bit desktop computers became common.  Most of the
changes here are to convert the test code to be 64-bit portable.

I have fixed this up in order to be able to use the fpu-tst test for
testing my new remainder function of the sim-fpu code.  That means that
this does depend on having my remainder patch as well [0].

The fpu-tst tool actually depends on external code from v2 of berkeley
softfloat [1] and testfloat [2] libraries which I cleaned up so the
would actually build on a modern compiler.  The build will generate a
'testfloat' binary which will be used to compare the sim-fpu vs the
softfloat implementation.

[0] https://sourceware.org/ml/gdb-patches/2017-10/msg00105.html
[1] https://github.com/stffrdhrn/berkeley-softfloat-2
[2] https://github.com/stffrdhrn/berkeley-testfloat-2

yyyy-mm-dd  Stafford Horne  <shorne@gmail.com>

	* common/Makefile.in: Allow out of tree make.  Fix fpu-tst
	targets.
	* common/alu-tst.c: Add includes to fix compile errors.
	* common/bits-gen.c: Add includes to fix compile errors.
	Changed 64 and 32 bit types to uit32_t and uint64_t to fix
	compiler and runtime errors. Fixed generated code defines and
	includes to fix compiler errors.
	* common/bits-tst.c: Add includes to fix compile errors.
	* common/fpu-tst.c: Add defines and includes to fix compile
	errors.
	(syst_float_flags_clear): New case for
	sim_fpu_status_invalid_irx.
	(syst_float_set_rounding_precision): New function.
	(syst_int64_to_float64): New function.
	(syst_int64_to_float32): New function.
	(syst_float64_to_int64_round_to_zero): New function.
	(syst_float32_to_int64_round_to_zero): New function.
	(syst_float32_rem): New function.
	(syst_float64_rem): New function.