or1k: offset needs to take into tcb alignment
The offset was wrong and causing issues when the TLS section alignment
is not 4 bytes.

i.e. in the thread section we have:

                 /- TP
                V
[ Pthread | TCB | TDATA , TBSS ]

The thread pointer (TP) points to the start of TDATA usually.  If the data is
aligned to 16 we may have.

                 /- TP
                V
[ Pthread | TCB 0x0 0x0 0x0 | TDATA , TBSS ]

Here TDATA is not at TP, but TDATA is offset by the alignment.  TP points
to just after the TCB.  With this patch I adjust for that TCB padding
due to alignment.  However, should we be patching GLIBC to set TP to
compensate for the alignment padding and have it point to TP? Or is this
the right thing to do?

Questions:
 - Why is TLS section alignment set to the highest aligned member?
 - Is this the right place to make the patch or should we be patching
   glibc?